Making it so that judges can have multiple affils.
This makes it so that the autogenerating scratches work with multiple
schools the judgeg is affiliated with.

IMPORTANT: This breaks the old schema, old DBs will fail quite hard
make sure that you use the new site.db and use an old version of mit-tab
to read your old databases.