User Interface Updates
Added a more intuitive filter UI to the Judge/Team/Debater lists.

Changed the scrolling in the pairing list to be activated by a toggle
button so that it is explicitly instead of implicitly happening. Also
made that view publicly available and sorted by gov team.

Re-worked the ballot entry page to be more like a ballot and less like a
giant list.