Add get/set to TabSettings model
This allows us to treat the TabSettings object like a dictionary,
getting and setting values to it backed by a database. This does not
change the fields of the table, so no migration is needed. Note that in
general using the cache is preferred to using this new interface, but if
you need the data persisted this is an easy way to do it.