Updated Forfeit Logic
Added support for forfeits with average speaks. Admin interface now has
an integer option ‘lenient_late.’ Debaters who are not checked in to
the tournament before this round will be dropped with average
speaks/ranks. Teams not checked in after this round will be dropped
with speaks of 0 and ranks of 3.5. Default value is 1 round