some changes
font change to the tab cards, reduced team and judge names to
alpha-numeric (so CSVs can be more easily printed), added easter-egg
(it’s in base.html with ‘monopolistic tabulation’), changed the thanks
template, did documentation for a number of documents.

one interesting bug. if you added a win-by-forfeit into the database
and then tried to look at the current pairings, something with the
cache would throw a module error relating to a static method. I
disabled the calls to the cache to make it work again. this means that
everything is currently calculated on the fly, which is probably bad.