Add ability to export as a Tournament Archive (#226)
The Tournament Archive or DebateXML is an XML-based format for the
interchange of debate tournament data. Its schema is currently under
review in TabbycatDebate/dta-spec#4.

This commit adds an admin form to export the site's data through that
specification for archival and analysis. The page is available through
'/archive/download/'.